Conflicts:
	activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb

+++ actionpack/test/template/url_helper_test.rb
@@ -20,9 +20,9 @@ class UrlHelperTest < ActiveSupport::TestCase
     get "/article/:id" => "foo#article", :as => :article
   end
 
+  include ActionView::Helpers::UrlHelper
   include routes.url_helpers
 
-  include ActionView::Helpers::UrlHelper
   include ActionView::Helpers::JavaScriptHelper
   include ActionDispatch::Assertions::DomAssertions
   include ActionView::Context
@@ -90,15 +90,33 @@ class UrlHelperTest < ActiveSupport::TestCase
   def test_button_to_with_javascript_confirm
     assert_dom_equal(
       "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      button_to("Hello", "http://www.example.com", :data => { :confirm => "Are you sure?" })
     )
+  end
+
+  def test_button_to_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      )
+    end
   end
 
   def test_button_to_with_javascript_disable_with
     assert_dom_equal(
       "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-disable-with=\"Greeting...\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", :disable_with => "Greeting...")
+      button_to("Hello", "http://www.example.com", :data => { :disable_with => "Greeting..." })
     )
+  end
+
+  def test_button_to_with_javascript_deprecated_disable_with
+    assert_deprecated ":disable_with option is deprecated and will be removed from Rails 4.1. Use ':data => { :disable_with => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-disable-with=\"Greeting...\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :disable_with => "Greeting...")
+      )
+    end
   end
 
   def test_button_to_with_remote_and_form_options
@@ -108,22 +126,33 @@ class UrlHelperTest < ActiveSupport::TestCase
   def test_button_to_with_remote_and_javascript_confirm
     assert_dom_equal(
       "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", :remote => true, :confirm => "Are you sure?")
+      button_to("Hello", "http://www.example.com", :remote => true, :data => { :confirm => "Are you sure?" })
     )
+  end
+
+  def test_button_to_with_remote_and_javascript_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :remote => true, :confirm => "Are you sure?")
+      )
+    end
   end
 
   def test_button_to_with_remote_and_javascript_disable_with
     assert_dom_equal(
       "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-disable-with=\"Greeting...\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", :remote => true, :disable_with => "Greeting...")
+      button_to("Hello", "http://www.example.com", :remote => true, :data => { :disable_with => "Greeting..." })
     )
   end
 
-  def test_button_to_with_remote_and_javascript_confirm_and_javascript_disable_with
-    assert_dom_equal(
-      "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-disable-with=\"Greeting...\" data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", :remote => true, :confirm => "Are you sure?", :disable_with => "Greeting...")
-    )
+  def test_button_to_with_remote_and_javascript_deprecated_disable_with
+    assert_deprecated ":disable_with option is deprecated and will be removed from Rails 4.1. Use ':data => { :disable_with => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-disable-with=\"Greeting...\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :remote => true, :disable_with => "Greeting...")
+      )
+    end
   end
 
   def test_button_to_with_remote_false
@@ -155,6 +184,13 @@ class UrlHelperTest < ActiveSupport::TestCase
     assert_dom_equal(
       "<form method=\"get\" action=\"http://www.example.com\" class=\"button_to\"><div><input type=\"submit\" value=\"Hello\" /></div></form>",
       button_to("Hello", "http://www.example.com", :method => :get)
+    )
+  end
+
+  def test_button_to_with_block
+    assert_dom_equal(
+      "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><button type=\"submit\"><span>Hello</span></button></div></form>",
+      button_to("http://www.example.com") { content_tag(:span, 'Hello') }
     )
   end
 
@@ -208,23 +244,44 @@ class UrlHelperTest < ActiveSupport::TestCase
 
   def test_link_tag_with_custom_onclick
     link = link_to("Hello", "http://www.example.com", :onclick => "alert('yay!')")
-    expected = %{<a href="http://www.example.com" onclick="alert('yay!')">Hello</a>}
+    expected = %{<a href="http://www.example.com" onclick="alert(&#39;yay!&#39;)">Hello</a>}
     assert_dom_equal expected, link
   end
 
   def test_link_tag_with_javascript_confirm
     assert_dom_equal(
       "<a href=\"http://www.example.com\" data-confirm=\"Are you sure?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      link_to("Hello", "http://www.example.com", :data => { :confirm => "Are you sure?" })
     )
     assert_dom_equal(
-      "<a href=\"http://www.example.com\" data-confirm=\"You can't possibly be sure, can you?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :confirm => "You can't possibly be sure, can you?")
+      "<a href=\"http://www.example.com\" data-confirm=\"You cant possibly be sure, can you?\">Hello</a>",
+      link_to("Hello", "http://www.example.com", :data => { :confirm => "You cant possibly be sure, can you?" })
     )
     assert_dom_equal(
-      "<a href=\"http://www.example.com\" data-confirm=\"You can't possibly be sure,\n can you?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :confirm => "You can't possibly be sure,\n can you?")
+      "<a href=\"http://www.example.com\" data-confirm=\"You cant possibly be sure,\n can you?\">Hello</a>",
+      link_to("Hello", "http://www.example.com", :data => { :confirm => "You cant possibly be sure,\n can you?" })
     )
+  end
+
+  def test_link_tag_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-confirm=\"Are you sure?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      )
+    end
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-confirm=\"You cant possibly be sure, can you?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :confirm => "You cant possibly be sure, can you?")
+      )
+    end
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-confirm=\"You cant possibly be sure,\n can you?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :confirm => "You cant possibly be sure,\n can you?")
+      )
+    end
   end
 
   def test_link_to_with_remote
@@ -272,16 +329,45 @@ class UrlHelperTest < ActiveSupport::TestCase
   def test_link_tag_using_post_javascript_and_confirm
     assert_dom_equal(
       "<a href=\"http://www.example.com\" data-method=\"post\" rel=\"nofollow\" data-confirm=\"Are you serious?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :method => :post, :confirm => "Are you serious?")
+      link_to("Hello", "http://www.example.com", :method => :post, :data => { :confirm => "Are you serious?" })
     )
+  end
+
+  def test_link_tag_using_post_javascript_and_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-method=\"post\" rel=\"nofollow\" data-confirm=\"Are you serious?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :method => :post, :confirm => "Are you serious?")
+      )
+    end
   end
 
   def test_link_tag_using_delete_javascript_and_href_and_confirm
     assert_dom_equal(
       "<a href='\#' rel=\"nofollow\" data-confirm=\"Are you serious?\" data-method=\"delete\">Destroy</a>",
-      link_to("Destroy", "http://www.example.com", :method => :delete, :href => '#', :confirm => "Are you serious?"),
+      link_to("Destroy", "http://www.example.com", :method => :delete, :href => '#', :data => { :confirm => "Are you serious?" }),
       "When specifying url, form should be generated with it, but not this.href"
     )
+  end
+
+  def test_link_tag_using_delete_javascript_and_href_and_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href='\#' rel=\"nofollow\" data-confirm=\"Are you serious?\" data-method=\"delete\">Destroy</a>",
+        link_to("Destroy", "http://www.example.com", :method => :delete, :href => '#', :confirm => "Are you serious?"),
+        "When specifying url, form should be generated with it, but not this.href"
+      )
+    end
+  end
+
+  def test_link_tag_with_block
+    assert_dom_equal '<a href="/"><span>Example site</span></a>',
+      link_to('/') { content_tag(:span, 'Example site') }
+  end
+
+  def test_link_tag_with_block_and_html_options
+    assert_dom_equal '<a class="special" href="/"><span>Example site</span></a>',
+      link_to('/', :class => "special") { content_tag(:span, 'Example site') }
   end
 
   def test_link_tag_using_block_in_erb
@@ -294,6 +380,16 @@ class UrlHelperTest < ActiveSupport::TestCase
       "<a href=\"/article/Gerd_M%C3%BCller\">Gerd Müller</a>",
       link_to("Gerd Müller", article_path("Gerd_Müller".html_safe))
     )
+  end
+
+  def test_link_tag_escapes_content
+    assert_dom_equal '<a href="/">Malicious &lt;script&gt;content&lt;/script&gt;</a>',
+      link_to("Malicious <script>content</script>", "/")
+  end
+
+  def test_link_tag_does_not_escape_html_safe_content
+    assert_dom_equal '<a href="/">Malicious <script>content</script></a>',
+      link_to("Malicious <script>content</script>".html_safe, "/")
   end
 
   def test_link_to_unless
